requirements.txt
click
pyyaml
snowflake-connector-python
